refactor to single wizard file and add installation
