+++ scripts/commands/wizard/utils.js
@@ -6,10 +6,78 @@ const log = require('npmlog')
 const marked = require('meta-marked')
 
 const tutorials = require('../../../src/static/tutorials.json')
+const courses = require('../../../src/static/courses.json')
+const projects = require('../../../src/static/projects.json')
 
 const tutorialKeys = Object.keys(tutorials)
 const latestTutorialId = tutorialKeys.sort().reverse()[0]
 const latestTutorial = tutorials[latestTutorialId]
+
+async function createTutorial () {
+  log.info("Let's create the files you need to build your tutorial. We'll ask you a few questions to get started.")
+  const responses = await inquirer
+    .prompt([
+      {
+        type: 'input',
+        name: 'title',
+        message: 'What is the name of your new tutorial?',
+        validate: validateStringPresent
+      },
+      {
+        type: 'input',
+        name: 'url',
+        message: 'What is the URL for your tutorial? (Hit return to accept our suggestion.)',
+        default: function (responses) {
+          return responses.title.toLowerCase().split(' ').join('-')
+        },
+        validate: validateStringPresent
+      },
+      {
+        type: 'list',
+        name: 'project',
+        message: 'Which project is your tutorial about?',
+        choices: projects.map(project => ({name: project.name, value: project.id}))
+      },
+      {
+        type: 'input',
+        name: 'description',
+        message: 'Please provide a short description for your tutorial to be displayed in tutorial listings.',
+        validate: validateStringPresent
+      }
+    ])
+
+  // determine new tutorial number
+  const tutorialNumber = nextTutorialNumber()
+
+  // create new directory
+  await promisify(fs.mkdir)(`src/tutorials/${tutorialNumber}-${responses.url}`)
+
+  // update all array in courses.json
+  courses.all.push(tutorialNumber)
+  await promisify(fs.writeFile)('src/static/courses.json', JSON.stringify(courses, null, 4))
+
+  // add entry to tutorials.json
+  const newTutorial = {
+    url: responses.url,
+    project: responses.project,
+    title: responses.title,
+    description: responses.description,
+    newMessage: '',
+    updateMessage: '',
+    createdAt: new Date(),
+    updatedAt: '',
+    resources: []
+  }
+
+  tutorials[tutorialNumber] = newTutorial
+  await promisify(fs.writeFile)('src/static/tutorials.json', JSON.stringify(tutorials, null, 4))
+
+  // log success
+  log.info(`Thanks! We've created a directory for your tutorial at \`src/tutorials/${tutorialNumber}-${responses.url}/\`.`)
+  log.info(`Preview your tutorial by running \`npm start\` and visiting: http://localhost:3000/#/${responses.url}`)
+  // suggest creating a lesson
+  await promptCreateFirst('lesson', tutorialNumber)
+}
 
 // used by TUTORIAL
 function nextTutorialNumber () {
@@ -119,6 +187,24 @@ function validateStringPresent (string) {
   }
 }
 
+async function createResourceIntro () {
+  // determine new tutorial number
+  log.info("Let's add resources to your tutorial.")
+  async function command (options) {
+    const { tutorial, tutorialId } = await selectTutorial('resources')
+    const resources = tutorials[tutorialId].resources
+
+    // print existing resources if present
+    if (resources.length === 0) {
+      log.info("Let's create your first resource!")
+    } else {
+      log.info("Here are the resources you've created so far:")
+      logResources(resources)
+    }
+
+    await createResource(tutorial, tutorialId) // loops until user declines to repeat, then offers closing statements
+  }
+}
 async function createResource (tutorial, tutorialId) {
   const responses = await inquirer
     .prompt([
@@ -164,6 +250,23 @@ async function createResource (tutorial, tutorialId) {
   await offerRepeat(tutorial, tutorialId, 'resource') // loops until user declines to repeat
 }
 
+async function createLessonIntro () {
+  // determine new tutorial number
+  log.info("Let's add lessons to your tutorial.")
+  const { tutorial, tutorialId, lessons } = await selectTutorial('lesson')
+
+  // print existing lessons if present
+  if (lessons.length === 0) {
+    log.info("Let's create your first lesson!")
+  } else {
+    log.info("Here are the lessons you've created so far:")
+    logLessons(lessons)
+    log.info("Let's create your next lesson!")
+  }
+
+  // loops until you say you don't want more lessons, then offers closing statements
+  await createLesson(tutorial, tutorialId, lessons)
+}
 // used by LESSONS
 async function createLesson (tutorial, tutorialId) {
   let lessons = await getTutorialLessons(tutorial, tutorialId)
@@ -324,4 +427,4 @@ async function promptCreateFirst (itemType, tutorialId) {
   }
 }
 
-module.exports = { getTutorialLessons, nextTutorialNumber, promptCreateFirst, offerRepeat, selectTutorial, validateStringPresent, nextLessonNumber, createLesson, createResource, logLessons, logResources }
+module.exports = { getTutorialLessons, createTutorial, createLessonIntro, createResourceIntro, nextTutorialNumber, promptCreateFirst, offerRepeat, selectTutorial, validateStringPresent, nextLessonNumber, createLesson, createResource, logLessons, logResources }
